Back-end/requirements.txt
# Backend Dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Optional for advanced features
# tensorflow==2.15.0
# keras==2.15.0
